feat: openrpc cmd using cobra, fixing nits
